requirements.txt
pyyaml
invoke
boto3
